add load language function
